arc_types::types

Struct ZeroizedBytes

pub struct ZeroizedBytes { /* private fields */ }

A secure byte container that zeroizes its contents on drop.

Security Note

Clone is intentionally NOT implemented to prevent creating copies of sensitive data that might not be properly zeroized. If you need to share the data, use as_slice() to get a reference.

Implementations

impl ZeroizedBytes

pub fn new(data: Vec<u8>) -> Self

Creates a new ZeroizedBytes from raw byte data.

pub fn as_slice(&self) -> &[u8]

Returns the data as a byte slice.

pub fn len(&self) -> usize

Returns the length of the data in bytes.

pub fn is_empty(&self) -> bool

Returns true if the data is empty.